+++ src/ibldsp/waveforms.py
@@ -11,7 +11,7 @@ import matplotlib as mpl
 import scipy
 
 import neuropixel
-from ibldsp.utils import parabolic_max, make_channel_index
+from ibldsp.utils import parabolic_max, make_channel_index, fcn_cosine
 from ibldsp.fourier import fshift
 
 EXTRACT_RADIUS_UM = 200
@@ -510,7 +510,7 @@ def recovery_point(arr_peak, df, idx_from_trough=5):
     # Check df['peak_time_idx'] + pt_idx is not out of bound
     idx_all = df["trough_time_idx"].to_numpy() + idx_from_trough
     # Find waveform(s) for which the second point is outside matrix boundary range
-    idx_over = np.where(idx_all > arr_peak.shape[1])[0]
+    idx_over = np.where(idx_all >= arr_peak.shape[1])[0]
     if len(idx_over) > 0:
         # Todo should this raise a warning ?
         idx_all[idx_over] = arr_peak.shape[1] - 1  # Take the last value of the waveform
@@ -652,6 +652,12 @@ def compute_spike_features(
     :return: dataframe of spikes with all features,
     Returns:
     """
+    arr_in = _validate_arr_in(arr_in)
+    n_taper = 5
+    n_samples = arr_in.shape[1]
+    t = np.arange(n_samples)
+    taper = fcn_cosine([0, n_taper])(t) * (1 - fcn_cosine([n_samples - n_taper, n_samples])(t))
+    arr_in = arr_in * taper[np.newaxis, :, np.newaxis]
     df = find_peak(arr_in)
     # Per waveform, keep only trace that contains the peak
     arr_peak_real = get_array_peak(arr_in, df)